chore: organize Schedule Page components a little
